* Sample read and playback
* Work on processors

+++ DrawingBoxes/Processing/Sample.cs
@@ -2,14 +2,14 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
+
 
 namespace VaporDAW
 {
     public struct Sample
     {
-        public float Left;
-        public float Right;
+        public double Left;
+        public double Right;
 
         public void Add(Sample other)
         {
